% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_search.R
\name{wos_search}
\alias{wos_search}
\title{Search Web of Science starter API}
\usage{
wos_search(query, filename = NULL, wos_base_folder = NULL, wos_api_key = NULL)
}
\arguments{
\item{query}{Web of Science query. See examples and official documentation.}

\item{filename}{Custom file name for caching the query, defaults to NULL. If
NULL, file name is derived from the query.}

\item{wos_base_folder}{Base folder where all responses from Web of Science
API calls are cached}

\item{wos_api_key}{Web of Science API.}
}
\value{
A list with all responses.
}
\description{
Search Web of Science starter API
}
\examples{
\dontrun{
if (interactive()) {
  # Remember to set the API key with wos_set()
  wos_search(query = "TS='higher education' AND FPY=2025")
}
}
}
