% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_set.R
\name{wos_set}
\alias{wos_set}
\title{Set basic options for calls to Web of Science API}
\usage{
wos_set(wos_base_folder = NULL, wos_api_key = NULL)
}
\arguments{
\item{wos_base_folder}{Base folder where all responses from Web of Science
API calls are cached}

\item{wos_api_key}{Web of Science API.}
}
\value{
Invisibly returns all settings as a list.
}
\description{
Set basic options for calls to Web of Science API
}
\examples{
wos_set(
  wos_base_folder = "WoS_cache",
  wos_api_key = "SuperSecretKey"
)

wos_settings <- wos_set()

wos_settings
}
